Move the Rust Scope code to its own file.
