+++ src/rust/engine/options/src/env.rs
@@ -6,10 +6,11 @@ use std::collections::HashMap;
 use std::env;
 use std::ffi::OsString;
 
-use super::id::{NameTransform, OptionId, Scope};
+use super::id::{NameTransform, OptionId};
 use super::{DictEdit, OptionsSource};
 use crate::fromfile::FromfileExpander;
 use crate::parse::Parseable;
+use crate::scope::Scope;
 use crate::ListEdit;
 
 #[derive(Debug)]
